google and facebook login

+++ project.py
@@ -1,20 +1,275 @@
-from flask import Flask, render_template, request, redirect,jsonify, url_for, flash
+import string
+import random
+import requests
+import json
+import httplib2
+from googleapiclient.discovery import build
+
+from google.auth.transport import requests as google_requests
+from google.oauth2 import id_token
+
+import google.oauth2.credentials
+
+
+from oauth2client.client import AccessTokenRefreshError
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+
+from flask import session as login_session
+from database_setup import Base, Restaurant, MenuItem, User
+from sqlalchemy.orm import sessionmaker
+from sqlalchemy import create_engine, asc
+from flask import Flask, render_template, request, redirect, jsonify, url_for, flash, make_response
 app = Flask(__name__)
 
-from sqlalchemy import create_engine, asc
-from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Restaurant, MenuItem
 
 
-#Connect to Database and create database session
-engine = create_engine('sqlite:///restaurantmenu.db')
+CLIENT_ID = json.loads(
+	open('client_secret.json', 'r').read())['web']['client_id']
+
+CLIENT_SECRETS_FILE = "client_secret.json"
+
+SCOPES = ['https://www.googleapis.com/auth/drive.metadata.readonly']
+API_SERVICE_NAME = 'drive'
+API_VERSION = 'v2'
+
+
+# Connect to Database and create database session
+engine = create_engine(
+    'sqlite:///restaurantmenu.db',
+    connect_args={
+        'check_same_thread': False},
+         echo=True)
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
+@app.route('/gconnect', methods=['GET', 'POST'])
+def gconnect():
+    print("Gconnect function")
+    if request.method == 'POST':
+        try:
+            if login_session['state'] != request.args.get('state'):
+                
+                response = make_response(json.dumps('Invalid state parameter.'), 401)
+                response.headers['Content-Type'] = 'application/json'
+                print("step 1")
+                return response
 
-#JSON APIs to view Restaurant Information
+			# Retrieve the token sent by the client
+            token = request.data
+            print("step 2")
+
+			# Request an access tocken from the google api
+            idinfo = id_token.verify_oauth2_token(token, google_requests.Request(), CLIENT_ID)
+            print("step 3")
+            url = (
+        	'https://oauth2.googleapis.com/tokeninfo?id_token=%s'
+                % token)
+            h = httplib2.Http()
+            result = json.loads(h.request(url, 'GET')[1])
+            print("step 4")
+            print(result['aud'])
+			# If there was an error in the access token info, abort.
+            if result.get('error') is not None:
+                response = make_response(json.dumps(result.get('error')), 500)
+                response.headers['Content-Type'] = 'application/json'
+                return response
+            print("step 5")
+			# Verify that the access token is used for the intended user.
+            user_google_id = idinfo['sub']
+            if result['sub'] != user_google_id:
+                response = make_response(
+                    json.dumps("Token's user ID doesn't match given user ID."),	401)
+                response.headers['Content-Type'] = 'application/json'
+                return response
+            print(result['sub'])
+			# Verify that the access token is valid for this app.
+            if result['aud'] != CLIENT_ID:
+                print("step 5.5")
+                response = make_response(json.dumps("Token's client ID does not match app's."), 401)
+                
+                print ("Token's client ID does not match app's.")
+                response.headers['Content-Type'] = 'application/json'
+                return response
+                        
+            print("step 6")
+			# Check if the user is already logged in
+            stored_access_token = login_session.get('access_token')
+            stored_user_google_id = login_session.get('user_google_id')
+            if stored_access_token is not None and user_google_id == stored_user_google_id:
+                response = make_response(json.dumps('Current user is already connected.'), 200)
+                response.headers['Content-Type'] = 'application/json'
+                return response
+            print("step 7")
+			# Store the access token in the session for later use.
+            login_session['access_token'] = idinfo
+            login_session['user_google_id'] = user_google_id
+			# Get user info
+            login_session['username'] = idinfo['name']
+            login_session['picture'] = idinfo['picture']
+            login_session['email'] = idinfo['email']
+            
+            user_id = getUserID(idinfo['email'])
+            
+            if not user_id:
+                user_id = createUser(login_session)
+            
+            login_session['user_id'] = user_id
+            
+            return 'Successful'
+			
+        except ValueError:
+			# Invalid token
+            print("value error")
+            pass
+
+
+@app.route('/disconnect', methods=['GET', 'POST'])
+def disconnect():
+  """Revoke current user's token and reset their session."""
+  # Only disconnect a connected user.
+  access_token = login_session.get('access_token')
+  
+  if access_token is None:
+    response = make_response(json.dumps('Current user not connected.'), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+
+  # Execute HTTP GET request to revoke current token.
+#   access_token = credentials.access_token
+  url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+  h = httplib2.Http()
+  result = h.request(url, 'GET')
+  
+#   del login_session['access_token']
+#   del login_session['gplus_id']
+#   del login_session['username']
+#   del login_session['email']
+#   del login_session['picture']
+
+  response = make_response(json.dumps('Successfully disconnected.'), 200)
+  response.headers['Content-Type'] = 'application/json'
+  
+
+  if result[0]['status'] == '200':
+    # Reset the user's session.
+    del login_session['access_token']
+    del login_session['gplus_id']
+    del login_session['username']
+    del login_session['email']
+    del login_session['picture']
+
+
+    response = make_response(json.dumps('Successfully disconnected.'), 200)
+    response.headers['Content-Type'] = 'application/json'
+    
+    return redirect(url_for('showRestaurants'))
+
+  else:
+    # For whatever reason, the given token was invalid.
+    response = make_response(
+        json.dumps('Failed to revoke token for given user.', 400))
+    response.headers['Content-Type'] = 'application/json'
+    return response
+
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits)
+                    for x in xrange(32))
+    login_session['state'] = state
+    # return "The current session state is %s" % login_session['state']
+    return render_template('login.html', STATE=state)
+
+@app.route('/welcome')
+def welcome():
+    return "welcome page"
+	# return render_template('welcome.html', picture=login_session['picture'], fullname=login_session['username'])
+
+@app.route('/fbconnect', methods=['POST'])
+def fbconnect():
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = request.data
+    print "access token received %s " % access_token
+
+
+    app_id = json.loads(open('fb_client_secrets.json', 'r').read())[
+        'web']['app_id']
+    app_secret = json.loads(
+        open('fb_client_secrets.json', 'r').read())['web']['app_secret']
+    url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (
+        app_id, app_secret, access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+
+
+    # Use token to get user info from API
+    userinfo_url = "https://graph.facebook.com/v2.8/me"
+    '''
+        Due to the formatting for the result from the server token exchange we have to
+        split the token first on commas and select the first index which gives us the key : value
+        for the server access token then we split it on colons to pull out the actual token value
+        and replace the remaining quotes with nothing so that it can be used directly in the graph
+        api calls
+    '''
+    token = result.split(',')[0].split(':')[1].replace('"', '')
+
+    url = 'https://graph.facebook.com/v2.8/me?access_token=%s&fields=name,id,email' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    # print "url sent for API access:%s"% url
+    # print "API JSON result: %s" % result
+    data = json.loads(result)
+    login_session['provider'] = 'facebook'
+    login_session['username'] = data["name"]
+    login_session['email'] = data["email"]
+    login_session['facebook_id'] = data["id"]
+
+    # The token must be stored in the login_session in order to properly logout
+    login_session['access_token'] = token
+
+    # Get user picture
+    url = 'https://graph.facebook.com/v2.8/me/picture?access_token=%s&redirect=0&height=200&width=200' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
+
+    login_session['picture'] = data["data"]["url"]
+
+    # see if user exists
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+
+    flash("Now logged in as %s" % login_session['username'])
+    return output
+
+
+@app.route('/fbdisconnect')
+def fbdisconnect():
+    facebook_id = login_session['facebook_id']
+    # The access token must me included to successfully logout
+    access_token = login_session['access_token']
+    url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id,access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'DELETE')[1]
+    return "you have been logged out"
+    
+# JSON APIs to view Restaurant Information
 @app.route('/restaurant/<int:restaurant_id>/menu/JSON')
 def restaurantMenuJSON(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
@@ -33,66 +288,84 @@ def restaurantsJSON():
     return jsonify(restaurants= [r.serialize for r in restaurants])
 
 
-#Show all restaurants
+# Show all restaurants
 @app.route('/')
 @app.route('/restaurant/')
 def showRestaurants():
   restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
+  print("****************************************")
+  if 'username' not in login_session:
+      print("no username in login session")
+      return render_template('publicrestaurants.html', restaurants=restaurants)
+    
+  print("user logged in")
+  print("user:", login_session['username'])
   return render_template('restaurants.html', restaurants = restaurants)
 
-#Create a new restaurant
+# Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
 def newRestaurant():
-  if request.method == 'POST':
-      newRestaurant = Restaurant(name = request.form['name'])
-      session.add(newRestaurant)
-      flash('New Restaurant %s Successfully Created' % newRestaurant.name)
-      session.commit()
-      return redirect(url_for('showRestaurants'))
-  else:
-      return render_template('newRestaurant.html')
+    if 'username' not in login_session:
+        return redirect('/login')
+    if request.method == 'POST':
+        newRestaurant = Restaurant(name = request.form['name'], user_id=login_session['user_id'])
+        session.add(newRestaurant)
+        flash('New Restaurant %s Successfully Created' % newRestaurant.name)
+        session.commit()
+        return redirect(url_for('showRestaurants'))
+    else:
+        return render_template('newRestaurant.html')
 
-#Edit a restaurant
+# Edit a restaurant
 @app.route('/restaurant/<int:restaurant_id>/edit/', methods = ['GET', 'POST'])
 def editRestaurant(restaurant_id):
-  editedRestaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
-  if request.method == 'POST':
-      if request.form['name']:
-        editedRestaurant.name = request.form['name']
-        flash('Restaurant Successfully Edited %s' % editedRestaurant.name)
-        return redirect(url_for('showRestaurants'))
-  else:
-    return render_template('editRestaurant.html', restaurant = editedRestaurant)
+    if 'username' not in login_session:
+            return redirect('/login')
+    editedRestaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    if request.method == 'POST':
+        if request.form['name']:
+            editedRestaurant.name = request.form['name']
+            flash('Restaurant Successfully Edited %s' % editedRestaurant.name)
+            return redirect(url_for('showRestaurants'))
+    else:
+        return render_template('editRestaurant.html', restaurant = editedRestaurant)
 
 
-#Delete a restaurant
+# Delete a restaurant
 @app.route('/restaurant/<int:restaurant_id>/delete/', methods = ['GET','POST'])
 def deleteRestaurant(restaurant_id):
-  restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
-  if request.method == 'POST':
-    session.delete(restaurantToDelete)
-    flash('%s Successfully Deleted' % restaurantToDelete.name)
-    session.commit()
-    return redirect(url_for('showRestaurants', restaurant_id = restaurant_id))
-  else:
-    return render_template('deleteRestaurant.html',restaurant = restaurantToDelete)
+    if 'username' not in login_session:
+            return redirect('/login')
+    restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    if request.method == 'POST':
+        session.delete(restaurantToDelete)
+        flash('%s Successfully Deleted' % restaurantToDelete.name)
+        session.commit()
+        return redirect(url_for('showRestaurants', restaurant_id = restaurant_id))
+    else:
+        return render_template('deleteRestaurant.html',restaurant = restaurantToDelete)
 
-#Show a restaurant menu
+# Show a restaurant menu
 @app.route('/restaurant/<int:restaurant_id>/')
 @app.route('/restaurant/<int:restaurant_id>/menu/')
 def showMenu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
+    if 'username' not in login_session:
+        return render_template('publicmenu.html', items = items, restaurant = restaurant)
     return render_template('menu.html', items = items, restaurant = restaurant)
      
 
 
-#Create a new menu item
+# Create a new menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/new/',methods=['GET','POST'])
 def newMenuItem(restaurant_id):
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
-      newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
+      newItem = MenuItem(name = request.form['name'], description = request.form['description'],
+        price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id, 
+        user_id=login_session['user_id'])
+
       session.add(newItem)
       session.commit()
       flash('New Menu %s Item Successfully Created' % (newItem.name))
@@ -100,7 +373,7 @@ def newMenuItem(restaurant_id):
   else:
       return render_template('newmenuitem.html', restaurant_id = restaurant_id)
 
-#Edit a menu item
+# Edit a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit', methods=['GET','POST'])
 def editMenuItem(restaurant_id, menu_id):
 
@@ -123,7 +396,7 @@ def editMenuItem(restaurant_id, menu_id):
         return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem)
 
 
-#Delete a menu item
+# Delete a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete', methods = ['GET','POST'])
 def deleteMenuItem(restaurant_id,menu_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
@@ -138,6 +411,26 @@ def deleteMenuItem(restaurant_id,menu_id):
 
 
 
+def createUser(login_session):
+    newUser = User(name=login_session['username'], email=login_session[
+                   'email'])
+                #    , picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
 
 if __name__ == '__main__':
   app.secret_key = 'super_secret_key'
